requirements.txt
langchain==0.3.25
langchain-groq==0.3.2
langsmith==0.3.42
python-dotenv==1.1.0
